tools/discord-bot/requirements.txt
discord.py>=2.3.0
python-dotenv
aiohttp
solana>=0.30.2
solders
